ai/requirements.txt
openai
python-dotenv
pyttsx3
sounddevice
scipy
numpy
requests
Pillow
pygame
beautifulsoup4